+++ rows/plugins/utils.py
@@ -28,45 +28,8 @@ if six.PY2:
 elif six.PY3:
     from collections.abc import Iterator
 
-from rows.fields import detect_types
+from rows.fields import detect_types, get_items, make_header, make_unique_name, slug, TextField
 from rows.table import FlexibleTable, Table
-
-SLUG_CHARS = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789_"
-
-
-def slug(text, separator="_", permitted_chars=SLUG_CHARS, replace_with_separator=" -_"):
-    """Generate a slug for the `text`.
-
-    >>> slug(' ÁLVARO  justen% ')
-    'alvaro_justen'
-    >>> slug(' ÁLVARO  justen% ', separator='-')
-    'alvaro-justen'
-    """
-
-    text = six.text_type(text or "")
-
-    # Strip non-ASCII characters
-    # Example: u' ÁLVARO  justen% ' -> ' ALVARO  justen% '
-    text = normalize("NFKD", text.strip()).encode("ascii", "ignore").decode("ascii")
-
-    # Replace spaces and other chars with separator
-    # Example: u' ALVARO  justen% ' -> u'_ALVARO__justen%_'
-    for char in replace_with_separator:
-        text = text.replace(char, separator)
-
-    # Remove non-permitted characters and put everything to lowercase
-    # Example: u'_ALVARO__justen%_' -> u'_alvaro__justen_'
-    text = "".join(char for char in text if char in permitted_chars).lower()
-
-    # Remove double occurrencies of separator
-    # Example: u'_alvaro__justen_' -> u'_alvaro_justen_'
-    double_separator = separator + separator
-    while double_separator in text:
-        text = text.replace(double_separator, separator)
-
-    # Strip separators
-    # Example: u'_alvaro_justen_' -> u'alvaro_justen'
-    return text.strip(separator)
 
 
 def ipartition(iterable, partition_size):
@@ -99,40 +62,6 @@ def get_filename_and_fobj(filename_or_fobj, mode="r", dont_open=False):
     return filename, fobj
 
 
-def make_unique_name(name, existing_names, name_format="{name}_{index}", start=2):
-    """Return a unique name based on `name_format` and `name`."""
-    index = start
-    new_name = name
-    while new_name in existing_names:
-        new_name = name_format.format(name=name, index=index)
-        index += 1
-
-    return new_name
-
-
-def make_header(field_names, permit_not=False):
-    """Return unique and slugged field names."""
-    slug_chars = SLUG_CHARS if not permit_not else SLUG_CHARS + "^"
-
-    header = [
-        slug(field_name, permitted_chars=slug_chars) for field_name in field_names
-    ]
-    result = []
-    for index, field_name in enumerate(header):
-        if not field_name:
-            field_name = "field_{}".format(index)
-        elif field_name[0].isdigit():
-            field_name = "field_{}".format(field_name)
-
-        if field_name in result:
-            field_name = make_unique_name(
-                name=field_name, existing_names=result, start=2
-            )
-        result.append(field_name)
-
-    return result
-
-
 def create_table(
     data,
     meta=None,
@@ -147,34 +76,58 @@ def create_table(
     """Create a rows.Table object based on data rows and some configurations
 
     - `skip_header` is only used if `fields` is set
-    - `samples` is only used if `fields` is `None`
+    - `samples` is only used if `fields` is `None`. If samples=None, all data
+      is filled in memory - use with caution.
     - `force_types` is only used if `fields` is `None`
     - `import_fields` can be used either if `fields` is set or not, the
       resulting fields will seek its order
     - `fields` must always be in the same order as the data
     """
-    # TODO: add warning if using skip_header and create skip_rows
-    #       (int, default = ?). Could be used if `fields` is set or not.
 
     table_rows = iter(data)
+    force_types = force_types or {}
+    if import_fields is not None:
+        import_fields = make_header(import_fields)
 
     if fields is None:  # autodetect field types
+        # TODO: may add `type_hints` parameter so autodetection can be easier
+        #       (plugins may specify some possible field types).
         header = make_header(next(table_rows))
 
         if samples is not None:
             sample_rows = list(islice(table_rows, 0, samples))
+            table_rows = chain(sample_rows, table_rows)
         else:
-            sample_rows = list(table_rows)
-        table_rows = chain(sample_rows, table_rows)
+            sample_rows = table_rows = list(table_rows)
 
-        # TODO: optimize field detection (ignore fields on `force_types` and
-        #       not in `import_fields`).
-        # TODO: add `type_hints` parameter so autodetection can be easier
-        #       (plugins may specify some possible field types).
-        fields = detect_types(header, sample_rows, *args, **kwargs)
+        # Detect field types using only the desired columns
+        detected_fields = detect_types(
+            header,
+            sample_rows,
+            skip_indexes=[
+                index for index, field in enumerate(header)
+                if field in force_types or field not in (import_fields or header)
+            ],
+            *args, **kwargs
+        )
+        # Check if any field was added during detecting process
+        new_fields = [
+            field_name
+            for field_name in detected_fields.keys()
+            if field_name not in header
+        ]
+        # Finally create the `fields` with both header and new field names,
+        # based on detected fields `and force_types`
+        fields = OrderedDict([
+            (field_name, detected_fields.get(field_name, TextField))
+            for field_name in header + new_fields
+        ])
+        fields.update(force_types)
 
-        if force_types is not None:
-            fields.update(force_types)
+        # Update `header` and `import_fields` based on new `fields`
+        header = list(fields.keys())
+        if import_fields is None:
+            import_fields = header
 
     else:  # using provided field types
         if not isinstance(fields, OrderedDict):
@@ -186,36 +139,27 @@ def create_table(
             _ = next(table_rows)
 
         header = make_header(list(fields.keys()))
+        if import_fields is None:
+            import_fields = header
 
         fields = OrderedDict(
             [(field_name, fields[key]) for field_name, key in zip(header, fields)]
         )
 
-    if import_fields is not None:
-        import_fields = make_header(import_fields)
+    diff = set(import_fields) - set(header)
+    if diff:
+        field_names = ", ".join('"{}"'.format(field) for field in diff)
+        raise ValueError("Invalid field names: {}".format(field_names))
+    fields = OrderedDict(
+        [(field_name, fields[field_name]) for field_name in import_fields]
+    )
 
-        diff = set(import_fields) - set(header)
-        if diff:
-            field_names = ", ".join('"{}"'.format(field) for field in diff)
-            raise ValueError("Invalid field names: {}".format(field_names))
-
-        fields = OrderedDict(
-            [(field_name, fields[field_name]) for field_name in import_fields]
-        )
-
-    fields_names_indexes = [
-        (field_name, header.index(field_name)) for field_name in fields.keys()
-    ]
-
-    # TODO: put this inside Table.__init__
+    get_row = get_items(*map(header.index, import_fields))
     table = Table(fields=fields, meta=meta)
-    for row in table_rows:
-        table.append(
-            {
-                field_name: row[field_index]
-                for field_name, field_index in fields_names_indexes
-            }
-        )
+    table.extend(
+        dict(zip(import_fields, get_row(row)))
+        for row in table_rows
+    )
 
     return table